+++ src/app/Balloon.App.Notification/.container.config.php
@@ -1,5 +1,7 @@
 <?php
 use Balloon\Hook;
+use Balloon\Migration;
+use Balloon\App\Notification\Migration\Delta\Installation;
 use Balloon\App\Notification\Hook\NewShareAdded;
 use Balloon\App\Notification\Hook\Subscription;
 use Balloon\App\Notification\Adapter\Db;
@@ -39,5 +41,13 @@ return [
                 'arguments' => ['object' => '{'.Http::class.'}']
             ],
         ]
+    ],
+    Migration::class => [
+        'calls' => [
+            Installation::class => [
+                'method' => 'injectDelta',
+                'arguments' => ['delta' => '{'.Installation::class.'}']
+            ],
+        ],
     ],
 ];